Create a module, add Common emitter entities
