+++ modules/id-emitter/src/main/scala/com/snowplowanalytics/snowplow/scalatracker/emitters/id/AsyncEmitter.scala
@@ -18,8 +18,6 @@ import scala.concurrent.ExecutionContext
 
 import com.snowplowanalytics.snowplow.scalatracker.Emitter._
 
-import RequestProcessor._
-
 object AsyncEmitter {
   // Avoid starting thread in constructor
   /**
