+++ modules/id-emitter/src/main/scala/com/snowplowanalytics/snowplow/scalatracker/emitters/id/RequestProcessor.scala
@@ -29,8 +29,6 @@ import com.snowplowanalytics.snowplow.scalatracker.Emitter._
 
 class RequestProcessor {
 
-  import RequestProcessor._
-
   // JSON object with Iglu URI to Schema for payload
   private val payloadBatchStub: JsonObject =
     JsonObject("schema" := Tracker.PayloadDataSchemaKey.toSchemaUri)
@@ -175,80 +173,5 @@ class RequestProcessor {
     case Success(s) if s.code >= 200 && s.code < 300 => CollectorSuccess(s.code)
     case Success(s)                                  => CollectorFailure(s.code)
     case Failure(e)                                  => TrackerFailure(e)
-  }
-
-}
-
-object RequestProcessor {
-
-  /** ADT for possible track results */
-  sealed trait CollectorResponse extends Product with Serializable
-
-  /** Success. Collector accepted an event */
-  final case class CollectorSuccess(code: Int) extends CollectorResponse
-
-  /** Collector refused an event. Probably wrong endpoint or outage */
-  final case class CollectorFailure(code: Int) extends CollectorResponse
-
-  /** Other failure. Timeout or network unavailability */
-  final case class TrackerFailure(throwable: Throwable) extends CollectorResponse
-
-  /** Emitter cannot continue retrying. Note that this is not *response* */
-  final case class RetriesExceeded(lastResponse: CollectorResponse) extends CollectorResponse
-
-  /** User-provided callback */
-  type Callback = (CollectorParams, CollectorRequest, CollectorResponse) => Unit
-
-  /** Payload (either GET or POST) ready to be send to collector */
-  sealed trait CollectorRequest extends Product with Serializable {
-
-    /** Attempt to send */
-    def attempt: Int
-
-    /** Check if emitters should keep sending this request */
-    def isFailed: Boolean = attempt >= 10
-
-    /** Increment attempt number. Must be used whenever payload failed */
-    def updateAttempt: CollectorRequest = this match {
-      case g: GetCollectorRequest  => g.copy(attempt = attempt + 1)
-      case p: PostCollectorRequest => p.copy(attempt = attempt + 1)
-    }
-
-    /**
-     * Return same payload, but with updated stm
-     * **Must** be used right before payload goes to collector
-     */
-    def updateStm: CollectorRequest = this match {
-      case GetCollectorRequest(_, map) =>
-        val stm = System.currentTimeMillis().toString
-        GetCollectorRequest(attempt, map.updated("stm", stm))
-      case PostCollectorRequest(_, list) =>
-        val stm = System.currentTimeMillis().toString
-        PostCollectorRequest(attempt, list.map(_.updated("stm", stm)))
-    }
-  }
-
-  /** Single event, supposed to passed with GET-request */
-  final case class GetCollectorRequest(attempt: Int, payload: EmitterPayload) extends CollectorRequest
-
-  /** Multiple events, supposed to passed with POST-request */
-  final case class PostCollectorRequest(attempt: Int, payload: List[EmitterPayload]) extends CollectorRequest
-
-  /** Collector preferences */
-  case class CollectorParams(host: String, port: Int, https: Boolean) {
-
-    /** Return stringified collector representation, e.g. `https://splw.acme.com:80/` */
-    def getUri: String = s"${if (https) "https" else "http"}://$host:$port"
-  }
-
-  object CollectorParams {
-
-    /** Construct collector preferences with correct default port */
-    def construct(host: String, port: Option[Int] = None, https: Boolean = false): CollectorParams =
-      port match {
-        case Some(p)       => CollectorParams(host, p, https)
-        case None if https => CollectorParams(host, 443, https = true)
-        case None          => CollectorParams(host, 80, https = false)
-      }
   }
 }
