+++ project/Dependencies.scala
@@ -20,6 +20,7 @@ object Dependencies {
     val igluCore   = "0.3.0"
     val circe      = "0.10.0"
     val catsEffect = "1.0.0"
+    val http4s     = "0.19.0-M1"
 
     // Scala (test only)
     val specs2     = "4.3.2"
@@ -33,6 +34,7 @@ object Dependencies {
     val igluCoreCirce  = "com.snowplowanalytics" %% "iglu-core-circe" % V.igluCore
     val circe          = "io.circe"              %% "circe-parser"    % V.circe
     val catsEffect     = "org.typelevel"         %% "cats-effect"     % V.catsEffect
+    val http4s         = "org.http4s"            %% "http4s-client"   % V.http4s
 
     // Scala (test only)
     val specs2        = "org.specs2"             %% "specs2-core"     % V.specs2      % "test"
